+++ NEWS.rst
@@ -16,6 +16,9 @@ tested on PyPy3.6.1 v7.1.1-beta0.
 
 **We have dropped support for Python 2 now.**
 
+``pairwise2`` now allows the input of parameters with keywords and returns the
+alignments as a list of ``namedtuples``.
+
 The codon tables have been updated to NCBI genetic code table version 4.5,
 which adds Cephalodiscidae mitochondrial as table 33.
 
@@ -39,6 +42,7 @@ possible, especially the following contributors:
 - Deepak Khatri
 - Hielke Walinga (first contribution)
 - Kai Blin
+- Markus Piotrowski
 - Peter Cock
 - Rob Miller
 - Sergio Valqui